finished setting up tailwind styles. started configuring backend to receive OTel trace data
