+++ frontend/src/components/AnnotationMenu/AnnotationMenu.tsx
@@ -1,47 +1,63 @@
-// Component to render the annotation menu
-import React, { useState, useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchAnnotations } from '../../features/annotation/annotationSlice';
-import { RootState } from '../../app/store';
-import { Annotation } from '../../types';
-import { ThunkDispatch } from 'redux-thunk';
-import { AnyAction } from 'redux';
+import React, { SetStateAction, Dispatch } from 'react';
+import { Fragment, useState } from 'react'
+import { Dialog, Transition } from '@headlessui/react'
+import { XMarkIcon } from '@heroicons/react/24/outline'
+import  AnnotationTable  from './AnnotationTable';
 
-export const AnnotationMenu = () => {
+type Props = {
+    open: boolean;
+    setOpen: Dispatch<SetStateAction<boolean>>;
+};
 
-  const dispatch: ThunkDispatch<RootState, unknown, AnyAction> = useDispatch();
-  const annotations = useSelector((state: RootState) => state.annotations)
-  // Fetch annotations when component mounts
-  useEffect(() => {
-      dispatch(fetchAnnotations());
-  }, [dispatch]);
-  console.log('annotations', annotations)
+export default function AnnotationMenu({ open, setOpen }: Props) {
 
-    return (
-      <div className="w-3/12 flex-none top-0 right-0 overflow-y-auto h-screen p-4 bg-slate-800">
-        <h2 className="text-xl text-sky-50 font-bold mb-4">Annotations</h2>
-        <table className="min-w-full divide-y divide-gray-200 shadow rounded-lg bg-gray-900">
-          <thead className="bg-gray-300">
-            <tr>
-              <th className="py-2 px-4">Node or Trace</th>
-              <th className="py-2 px-4">Date</th>
-              <th className="py-2 px-4">Title</th>
-              <th className="py-2 px-4">Body</th>
-            </tr>
-          </thead>
-          <tbody className="bg-white divide-y divide-gray-200">
-            {annotations.list.map((annotation: Annotation) => (
-              <tr key={annotation.annotationDate}>
-                <td align="left" className="py-2 px-4">
-                    {annotation.nodeId ? "Node" : "Trace"} {/* You can modify this based on your actual data structure */}
-                </td>
-                <td align="left" className="py-2 px-4">{annotation.annotationDate}</td>
-                <td align="right" className="py-2 px-4">{annotation.annotationName}</td>
-                <td align="right" className="py-2 px-4">{annotation.annotationBody}</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    );
-  }
+  return (
+    <Transition.Root show={open} as={Fragment}>
+      <Dialog as="div" className="relative z-10" onClose={setOpen}>
+        <div className="fixed inset-0" />
+
+        <div className="fixed inset-0 overflow-hidden">
+          <div className="absolute inset-0 overflow-hidden">
+            <div className="pointer-events-none fixed inset-y-0 right-0 flex max-w-full pl-10">
+              <Transition.Child
+                as={Fragment}
+                enter="transform transition ease-in-out duration-500 sm:duration-700"
+                enterFrom="translate-x-full"
+                enterTo="translate-x-0"
+                leave="transform transition ease-in-out duration-500 sm:duration-700"
+                leaveFrom="translate-x-0"
+                leaveTo="translate-x-full"
+              >
+                <Dialog.Panel className="pointer-events-auto w-screen max-w-md">
+                  <div className="flex h-full flex-col overflow-y-scroll bg-white py-6 shadow-xl">
+                    <div className="px-4 sm:px-6">
+                      <div className="flex items-start justify-between">
+                        <Dialog.Title className="text-base font-semibold leading-6 text-gray-900">
+                          Saved Annotations
+                        </Dialog.Title>
+                        <div className="ml-3 flex h-7 items-center">
+                          <button
+                            type="button"
+                            className="relative rounded-md bg-white text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
+                            onClick={() => setOpen(false)}
+                          >
+                            <span className="absolute -inset-2.5" />
+                            <span className="sr-only">Close panel</span>
+                            <XMarkIcon className="h-6 w-6" aria-hidden="true" />
+                          </button>
+                        </div>
+                      </div>
+                    </div>
+                    <div className="relative mt-6 flex-1 px-4 sm:px-6">
+                        <AnnotationTable />
+                    </div>
+                  </div>
+                </Dialog.Panel>
+              </Transition.Child>
+            </div>
+          </div>
+        </div>
+      </Dialog>
+    </Transition.Root>
+  )
+}
