+++ frontend/src/components/AnnotationMenu/AnnotationTable.tsx
