+++ frontend/src/features/index.js
@@ -1,0 +1,4 @@
+
+
+ghostVariable.ghostAction();
+
