+++ package.json
@@ -16,6 +16,7 @@
     "@babel/core": "^7.23.2",
     "@babel/preset-env": "^7.23.2",
     "@fontsource/roboto": "^5.0.8",
+    "@headlessui/react": "^1.7.17",
     "@reduxjs/toolkit": "^1.9.7",
     "@types/pg": "^8.10.7",
     "@types/react": "^18.2.31",
@@ -39,6 +40,7 @@
     "react-router-dom": "^6.17.0",
     "router-dom": "^2.2.11",
     "style-loader": "^3.3.3",
+    "tailwindcss": "^3.3.5",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2",
     "webpack": "^5.89.0",
